+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 * (breaking) Drop gradle 5 and 6 support
 * (breaking) Drop java 8 support (java 11 required)
 * All tasks properties replaced with lazy properties
+* Remove git-publish plugin. Plugin functions added directly with required modifications
+  so usage will be the same (compatible). The Author of the original plugin deprecated it
+  (together with grgit). Still, grgit will be used for now (because of its feature completeness)
 
 ### 3.0.0 (2022-11-01)
 * (breaking) Drop gradle 5.0-5.2 support (minimum required gradle is 5.3)